requirements.txt
argh==0.26.2
bat==0.3.4
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
pandas==0.24.2
pathtools==0.1.2
pkg-resources==0.0.0
py4j==0.10.7
pyarrow==0.13.0
pyparsing==2.4.0
pyspark==2.4.1
python-dateutil==2.8.0
python-geoip==1.2
python-geoip-geolite2==2015.303
python-geoip-python3==1.3
pytz==2019.1
PyYAML==5.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
urllib3==1.24.1
watchdog==0.9.0
